fix: documentation/testing/CVM
