+++ lintit.sh
@@ -7,9 +7,32 @@ fi
 
 PY_FILES=`find $BASE_PATH -name "*.py" -not -path "*/site-packages/*"`
 
-echo "yapf formatting"
+echo "yapf checking"
 
 for filename in $PY_FILES
 do
-   yapf --in-place $filename --style ./setup.cfg 
+   yapf -d $filename --style ./setup.cfg
+done
+
+
+echo "mypy check"
+
+for filename in $PY_FILES
+do
+    mypy $filename --config-file ./setup.cfg
+done
+
+
+echo "flake8 check"
+
+for filename in $PY_FILES
+do
+    flake8 $filename --config ./setup.cfg
+done
+
+
+echo "pylint check"
+for filename in $PY_FILES
+do
+    pylint $filename --rcfile ./setup.cfg
 done
