+++ setup.cfg
@@ -17,8 +17,9 @@ max-complexity = 10
 
 
 [pycodestyle]
-max-line-length = 123
+max-line-length = 120
 
 [pylint]
 disable = too-few-public-methods
 extension-pkg-whitelist=pygame
+max-line-length = 120
